added some examples for basic functions, created machines module
